requirements.txt
torch>=2.0.0
huggingface-hub>=0.17.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
